feat: add code for HTTP server
